Refactor to further split protocol and HA logic
